Modify AbstractShardingJDBCDatabaseAndTableTest, reuse datasource and close datasource after test case finished.
