+++ sharding-jdbc-core/src/test/java/io/shardingjdbc/core/integrate/type/sharding/hint/base/AbstractHintTest.java
@@ -45,10 +45,6 @@ import static org.dbunit.Assertion.assertEquals;
 
 public abstract class AbstractHintTest extends AbstractSQLTest {
     
-    private static boolean isShutdown;
-    
-    private static Map<DatabaseType, ShardingDataSource> shardingDataSources = new HashMap<>();
-    
     public AbstractHintTest() {
         
     }
@@ -78,30 +74,16 @@ public abstract class AbstractHintTest extends AbstractSQLTest {
                 "integrate/dataset/sharding/db/init/db_9.xml");
     }
     
-    protected final Map<DatabaseType, ShardingDataSource> getShardingDataSources() throws SQLException {
-        if (!shardingDataSources.isEmpty() && !isShutdown) {
-            return shardingDataSources;
-        }
-        isShutdown = false;
+    protected final Map<DatabaseType, ShardingDataSource> initShardingDataSources() throws SQLException {
         Map<DatabaseType, Map<String, DataSource>> dataSourceMap = createDataSourceMap();
         for (Map.Entry<DatabaseType, Map<String, DataSource>> each : dataSourceMap.entrySet()) {
             ShardingRule shardingRule = getShardingRule(each);
-            shardingDataSources.put(each.getKey(), new ShardingDataSource(shardingRule));
+            getShardingDataSources().put(each.getKey(), new ShardingDataSource(shardingRule));
         }
-        return shardingDataSources;
+        return getShardingDataSources();
     }
     
     protected abstract ShardingRule getShardingRule(Map.Entry<DatabaseType, Map<String, DataSource>> dataSourceEntry) throws SQLException;
-    
-    @AfterClass
-    public static void clear() {
-        isShutdown = true;
-        if (!shardingDataSources.isEmpty()) {
-            for (ShardingDataSource each : shardingDataSources.values()) {
-                each.close();
-            }
-        }
-    }
     
     protected void assertDataSet(final String expectedDataSetFile, final HintShardingValueHelper helper, 
                                  final Connection connection, final String sql,
