+++ sharding-jdbc-core/src/test/java/io/shardingjdbc/core/integrate/type/sharding/hint/type/RoutingDatabaseOnlyWithHintForDMLTest.java
@@ -44,7 +44,7 @@ public class RoutingDatabaseOnlyWithHintForDMLTest extends AbstractRoutingDataba
     
     @Before
     public void init() throws SQLException {
-        shardingDataSources = getShardingDataSources();
+        shardingDataSources = initShardingDataSources();
     }
     
     @Test
