+++ sharding-jdbc-core/src/test/java/io/shardingjdbc/core/integrate/type/sharding/hint/type/RoutingDatabaseOnlyWithHintForSelectTest.java
@@ -37,7 +37,7 @@ public class RoutingDatabaseOnlyWithHintForSelectTest extends AbstractRoutingDat
     
     @Before
     public void init() throws SQLException {
-        shardingDataSources = getShardingDataSources();
+        shardingDataSources = initShardingDataSources();
     }
     
     @Test
