+++ sharding-jdbc-core/src/test/java/io/shardingjdbc/core/integrate/type/sharding/hint/type/ShardingDatabaseOnlyWithHintForSelectTest.java
@@ -39,7 +39,7 @@ public final class ShardingDatabaseOnlyWithHintForSelectTest extends AbstractSha
     
     @Before
     public void init() throws SQLException {
-        shardingDataSources = getShardingDataSources();
+        shardingDataSources = initShardingDataSources();
     }
     
     @Test
